api/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
compositeai==0.1.9
distro==1.9.0
fastapi==0.115.2
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.20.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
openai==1.52.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
starlette==0.40.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
